experiments/6-finance-correlations/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gcp-real-time-market-predictor"
version = "0.1.0"
description = "GCP Real-Time Market Predictor: real-time sensor-driven market direction predictions with Dash UI"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "GCP2", email = "devnull@example.com" }
]
dependencies = [
  "dash==2.17.1",
  "dash-bootstrap-components==1.6.0",
  "plotly==5.23.0",
  "requests==2.32.3",
  "numpy==2.0.1",
  "PyYAML==6.0.2",
  "python-dotenv==1.0.1",
  "pydantic==2.8.2",
  "pydantic-settings==2.4.0",
  "typer==0.12.3",
  "tenacity==9.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest==8.3.2",
  "mypy==1.11.1",
  "ruff==0.5.7",
  "types-requests==2.32.0.20240712",
  "types-python-dateutil==2.9.0.20240316",
]

market = [
  "alpaca-trade-api==3.1.1",
  "pandas==2.2.2",
]

[project.scripts]
gcp-market = "main:app"

[tool.setuptools]
py-modules = ["main"]

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["experiments*", "tests*", "bq_cache*"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "UP", "B", "S", "W", "C90"]
ignore = ["S101", "S311"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_unused_configs = true
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true




